project_analyzer/requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
networkx==3.2.1
